[#106] refactor: auth 폴더와 jwt 폴더 구분 및 이동
