+++ src/main/java/com/jeju/nanaland/domain/favorite/controller/FavoriteController.java
@@ -8,8 +8,8 @@ import com.jeju.nanaland.domain.favorite.dto.FavoriteResponse.StatusDto;
 import com.jeju.nanaland.domain.favorite.service.FavoriteService;
 import com.jeju.nanaland.domain.member.dto.MemberResponse.MemberInfoDto;
 import com.jeju.nanaland.global.BaseResponse;
+import com.jeju.nanaland.global.auth.AuthMember;
 import com.jeju.nanaland.global.exception.SuccessCode;
-import com.jeju.nanaland.global.jwt.AuthMember;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
