+++ src/main/java/com/jeju/nanaland/domain/member/controller/MemberController.java
@@ -12,8 +12,8 @@ import com.jeju.nanaland.domain.member.dto.MemberResponse.RecommendPostDto;
 import com.jeju.nanaland.domain.member.service.MemberLoginService;
 import com.jeju.nanaland.domain.member.service.MemberTypeService;
 import com.jeju.nanaland.global.BaseResponse;
-import com.jeju.nanaland.global.jwt.AuthMember;
-import com.jeju.nanaland.global.jwt.dto.JwtResponseDto.JwtDto;
+import com.jeju.nanaland.global.auth.AuthMember;
+import com.jeju.nanaland.global.auth.jwt.dto.JwtResponseDto.JwtDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
