+++ src/main/java/com/jeju/nanaland/domain/nana/controller/NanaController.java
@@ -6,8 +6,8 @@ import com.jeju.nanaland.domain.nana.dto.NanaResponse.NanaDetailDto;
 import com.jeju.nanaland.domain.nana.dto.NanaResponse.NanaThumbnailDto;
 import com.jeju.nanaland.domain.nana.service.NanaService;
 import com.jeju.nanaland.global.BaseResponse;
+import com.jeju.nanaland.global.auth.AuthMember;
 import com.jeju.nanaland.global.exception.SuccessCode;
-import com.jeju.nanaland.global.jwt.AuthMember;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
