+++ src/main/java/com/jeju/nanaland/domain/search/controller/SearchController.java
@@ -8,8 +8,8 @@ import com.jeju.nanaland.domain.search.dto.SearchResponse.CategoryDto;
 import com.jeju.nanaland.domain.search.dto.SearchResponse.SearchVolumeDto;
 import com.jeju.nanaland.domain.search.service.SearchService;
 import com.jeju.nanaland.global.BaseResponse;
+import com.jeju.nanaland.global.auth.AuthMember;
 import com.jeju.nanaland.global.exception.SuccessCode;
-import com.jeju.nanaland.global.jwt.AuthMember;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
