+++ src/main/java/com/jeju/nanaland/global/auth/AuthMember.java
@@ -1,4 +1,4 @@
-package com.jeju.nanaland.global.jwt;
+package com.jeju.nanaland.global.auth;
 
 import io.swagger.v3.oas.annotations.Parameter;
 import java.lang.annotation.ElementType;
