+++ src/main/java/com/jeju/nanaland/global/auth/AuthMemberArgumentResolver.java
@@ -1,7 +1,8 @@
-package com.jeju.nanaland.global.jwt;
+package com.jeju.nanaland.global.auth;
 
 import com.jeju.nanaland.domain.member.dto.MemberResponse.MemberInfoDto;
 import com.jeju.nanaland.domain.member.repository.MemberRepository;
+import com.jeju.nanaland.global.auth.jwt.JwtUtil;
 import com.jeju.nanaland.global.exception.ErrorCode;
 import com.jeju.nanaland.global.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
