+++ src/main/java/com/jeju/nanaland/global/auth/jwt/dto/JwtResponseDto.java
@@ -1,4 +1,4 @@
-package com.jeju.nanaland.global.jwt.dto;
+package com.jeju.nanaland.global.auth.jwt.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
