+++ src/main/java/com/jeju/nanaland/global/config/SecurityConfig.java
@@ -1,10 +1,10 @@
 package com.jeju.nanaland.global.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jeju.nanaland.global.jwt.JWTAuthenticationFilter;
-import com.jeju.nanaland.global.jwt.JwtUtil;
-import com.jeju.nanaland.global.jwt.handler.CustomAccessDeniedHandler;
-import com.jeju.nanaland.global.jwt.handler.CustomAuthenticationEntryPoint;
+import com.jeju.nanaland.global.auth.handler.CustomAccessDeniedHandler;
+import com.jeju.nanaland.global.auth.handler.CustomAuthenticationEntryPoint;
+import com.jeju.nanaland.global.auth.jwt.JWTAuthenticationFilter;
+import com.jeju.nanaland.global.auth.jwt.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
